Refactor: util 함수들 패키지 옮기기
