+++ app-server/subprojects/script/src/main/kotlin/club/staircrusher/data_restore/AccessibilityInserter.kt
@@ -7,7 +7,10 @@ import club.staircrusher.place.application.port.out.web.MapsService
 import club.staircrusher.place.domain.model.Building
 import club.staircrusher.place.infra.adapter.out.web.KakaoMapsService
 import club.staircrusher.place.infra.adapter.out.web.KakaoProperties
+import club.staircrusher.readTsvAsLines
 import club.staircrusher.stdlib.domain.entity.EntityIdGenerator
+import club.staircrusher.toKstString
+import club.staircrusher.writeTextAsFile
 import kotlinx.coroutines.runBlocking
 import java.time.Instant
