+++ app-server/subprojects/script/src/main/kotlin/club/staircrusher/place_count/NearbyPlaceCounter.kt
@@ -1,6 +1,6 @@
 package club.staircrusher.place_count
 
-import club.staircrusher.data_restore.readTsvAsLines
+import club.staircrusher.readTsvAsLines
 import club.staircrusher.place.application.port.out.web.MapsService
 import club.staircrusher.place.infra.adapter.out.web.KakaoMapsService
 import club.staircrusher.place.infra.adapter.out.web.KakaoProperties
